fix: fix ios ComposeApp module not found

+++ shared/data/src/commonMain/kotlin/com/xiaoyv/bangumi/shared/data/model/response/image/ComposeAnimePicture.kt
@@ -54,7 +54,7 @@ data class ComposeAnimePictureImage(
     @SerialName("pubtime")
     val pubtime: String? = null,
     @SerialName("score")
-    val score: Int = 0,
+    val score: Float = 0f,
     @SerialName("score_number")
     val scoreNumber: Int = 0,
     @SerialName("size")
